+++ react-command-palette/tailwind.config.js
@@ -1,6 +1,15 @@
+import defaultTheme from "tailwindcss/defaultTheme";
+
 /** @type {import('tailwindcss').Config} */
 export default {
   content: ["./index.html", "./*.{ts,tsx}"],
-  theme: { extend: {} },
+  theme: {
+    extend: {
+      fontFamily: {
+        sans: ['"Instrument Sans"', ...defaultTheme.fontFamily.sans],
+        mono: ['"Martian Mono"', ...defaultTheme.fontFamily.mono],
+      },
+    },
+  },
   plugins: [],
 };